# Conflicts:
#	YumYum/src/main/java/yumyum/demo/src/community/controller/CommunityController.java
#	YumYum/src/main/java/yumyum/demo/src/community/repository/PostRepository.java
#	YumYum/src/main/java/yumyum/demo/src/community/service/CommunityService.java
#	YumYum/src/main/java/yumyum/demo/src/restaurant/entity/SearchEntity.java
#	YumYum/src/main/java/yumyum/demo/src/restaurant/repository/SearchRepository.java
#	YumYum/src/main/java/yumyum/demo/src/restaurant/service/RestaurantService.java

+++ YumYum/src/main/java/yumyum/demo/src/community/dto/CommentInfoDto.java
@@ -18,7 +18,7 @@ public class CommentInfoDto {
     private String nickName;
     private String profileImgUrl;
     private String contents;
-    private Long countCommentLike;
+    private Integer countCommentLike;
     private String createdAt;
     private int groupNumber;
     private Long parentId;
